+++ cmd/mailchain/internal/config/sent_store.go
@@ -19,7 +19,19 @@ import (
 	"github.com/mailchain/mailchain/stores"
 	"github.com/mailchain/mailchain/stores/s3store"
 	"github.com/pkg/errors"
+	"github.com/spf13/viper"
 )
+
+//go:generate mockgen -source=sent_store.go -package=configtest -destination=./configtest/sent_store_mock.go
+
+type SentStoreSetter interface {
+	Set(sentType string) error
+}
+
+type SentStore struct {
+	viper         *viper.Viper
+	requiredInput func(label string) (string, error)
+}
 
 // GetSentStore create all the clients based on configuration
 func (s SentStore) Get() (stores.Sent, error) {
