+++ cmd/mailchain/internal/setup/storage_sent.go
@@ -17,16 +17,8 @@ package setup
 import (
 	"github.com/mailchain/mailchain/cmd/mailchain/internal/config"
 	"github.com/mailchain/mailchain/cmd/mailchain/internal/config/names"
-	"github.com/mailchain/mailchain/cmd/mailchain/internal/prompts"
 	"github.com/spf13/viper" // nolint: depguard
 )
-
-func DefaultSentStorage() SentStorage {
-	return SentStorage{
-		sentStoreSetter:    config.DefaultSentStore(),
-		selectItemSkipable: prompts.SelectItemSkipable,
-	}
-}
 
 type SentStorage struct {
 	sentStoreSetter    config.SentStoreSetter
